Split debos main file
Initially 'debos.go' contained a lot of additions and logical parts in
single file. So move logical parts from 'debos.go' to appropriate packages:
- 'debos' package: actions interface and filesystem addons
- 'recipe' package: logic of Yaml-based recipe parsing

Use correct imports, naming and syntax for splitted files.
